spi2pcap.py: Extract packet timestamps from the SPI capture
This change extracts the timestamps between Ethernet
packets. The base time for the capture is taken as the time of
conversion as there is no absolute time information in the
CSV output from Saleae Logic. The input file timestamp could
be used but this will just reflect the checkout time from git.

Tests:
 - Convert the existing SPI captures and verify start and end
   times in the SPI capture are reflected in Wireshark
 - Verify that the file sizes of freshly converted pcap are
   identical to the existing ones.
 - Visually compare two captures side-by-side in Wireshark
   and verify timestamps are the only difference and that
   packet ordering is identical