spi2pcap.py: Extend to support Saleae Logic 2.x format input
Saleae Logic 2.x SPI decodes produce a slightly different form
of CSV decode output. This changes allows the script to
automatically detect the incoming data and decide which
columns contain the data that it needs. In theory it should be
extensible to input from other tools (e.g. Sigrok) in the future.

Small refactorings to move all of the main program to the
end of the file. Also simplify if statements to remove
extraneous brackets as this isn't C...

Tests:
 - Bulk convert existing Saleae Logic 1.x and new 2.x captures.
   Verify existing captures are unchanged and new captures
   appear to make sense.